patch and delete functions, see books.route.peter.js
